Driver: totalDistance to float + add to PUT

+++ src/main/java/com/vms/demo/dto/driver/DriverDTO.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class DriverDTO {
     private Long userID;
     private String drivingLicense;
-    private int totalDistance;
+    private float totalDistance;
     private Long totalTime;
     private int jobsDone;
 }
